make former unit test to still be using the embeeded old elasticsearch version and add new integration test based on testcontainers with a simple initial example on mappings
