--- src/test/java/io/confluent/connect/elasticsearch/TestContainerTest.java
@@ -1,218 +1,0 @@
-/*
- * Copyright 2018 Confluent Inc.
- *
- * Licensed under the Confluent Community License (the "License"); you may not use
- * this file except in compliance with the License.  You may obtain a copy of the
- * License at
- *
- * http://www.confluent.io/confluent-community-license
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
- * WARRANTIES OF ANY KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations under the License.
- */
-
-package io.confluent.connect.elasticsearch;
-
-import static io.confluent.connect.elasticsearch.DataConverter.BehaviorOnNullValues;
-import static io.confluent.connect.elasticsearch.ElasticsearchSinkConnectorConstants.KEYWORD_TYPE;
-import static io.confluent.connect.elasticsearch.ElasticsearchSinkConnectorConstants.TEXT_TYPE;
-import static junit.framework.TestCase.assertNotNull;
-import static junit.framework.TestCase.assertNull;
-import static org.junit.Assert.assertEquals;
-import static org.mockito.Mockito.mock;
-import static org.mockito.Mockito.when;
-
-import com.fasterxml.jackson.databind.JsonNode;
-import com.fasterxml.jackson.databind.node.NumericNode;
-import com.fasterxml.jackson.databind.node.ObjectNode;
-import com.fasterxml.jackson.databind.node.TextNode;
-import com.google.gson.JsonObject;
-import io.confluent.connect.elasticsearch.jest.JestElasticsearchClient;
-import java.util.Collections;
-import org.apache.kafka.connect.data.Date;
-import org.apache.kafka.connect.data.Decimal;
-import org.apache.kafka.connect.data.Field;
-import org.apache.kafka.connect.data.Schema;
-import org.apache.kafka.connect.data.SchemaBuilder;
-import org.apache.kafka.connect.data.Time;
-import org.apache.kafka.connect.data.Timestamp;
-import org.junit.After;
-import org.junit.Before;
-import org.junit.Test;
-import org.testcontainers.elasticsearch.ElasticsearchContainer;
-
-public class TestContainerTest {
-
-  private static final String INDEX = "kafka-connect";
-  private static final String TYPE = "kafka-connect-type";
-
-
-  private ElasticsearchContainer container;
-
-  @Before
-  public void setUp() {
-    container = new ElasticsearchContainer("docker.elastic.co/elasticsearch/elasticsearch:6.4.1");
-    container.start();
-
-    client = new JestElasticsearchClient("http://"+container.getHttpHostAddress());
-    converter = new DataConverter(true, BehaviorOnNullValues.IGNORE);
-  }
-
-  @After
-  public void tearDown() {
-    if (client != null) {
-      client.close();
-    }
-    client = null;
-    container.stop();
-  }
-
-
-  protected ElasticsearchClient client;
-  private DataConverter converter;
-
-  @Test
-  @SuppressWarnings("unchecked")
-  public void testMapping() throws Exception {
-
-    client.createIndices(Collections.singleton(INDEX));
-    Schema schema = createSchema();
-    Mapping.createMapping(client, INDEX, TYPE, schema);
-
-    JsonObject mapping = Mapping.getMapping(client, INDEX, TYPE);
-    assertNotNull(mapping);
-    verifyMapping(schema, mapping);
-  }
-
-  @Test
-  @SuppressWarnings("unchecked")
-  public void testStringMappingForES6() throws Exception {
-    ElasticsearchClient client = mock(ElasticsearchClient.class);
-    when(client.getVersion()).thenReturn(ElasticsearchClient.Version.ES_V6);
-
-    Schema schema = SchemaBuilder.struct().name("textRecord")
-        .field("string", Schema.STRING_SCHEMA)
-        .build();
-    ObjectNode mapping = (ObjectNode) Mapping.inferMapping(client, schema);
-    ObjectNode properties = mapping.with("properties");
-    ObjectNode string = properties.with("string");
-    TextNode stringType = (TextNode) string.get("type");
-    ObjectNode fields = string.with("fields");
-    ObjectNode keyword = fields.with("keyword");
-    TextNode keywordType = (TextNode) keyword.get("type");
-    NumericNode ignoreAbove = (NumericNode) keyword.get("ignore_above");
-
-    assertEquals(TEXT_TYPE, stringType.asText());
-    assertEquals(KEYWORD_TYPE, keywordType.asText());
-    assertEquals(256, ignoreAbove.asInt());
-  }
-
-  @Test
-  public void testInferMapping() throws Exception {
-
-    Schema stringSchema = SchemaBuilder
-        .struct()
-        .name("record")
-        .field("foo", SchemaBuilder.string().defaultValue("0").build())
-        .build();
-    JsonNode stringMapping = Mapping.inferMapping(client, stringSchema);
-
-    assertNull(stringMapping.get("properties").get("foo").get("null_value"));
-
-    Schema intSchema =SchemaBuilder
-        .struct()
-        .name("record")
-        .field("foo", SchemaBuilder.int32().defaultValue(0).build())
-        .build();
-
-    JsonNode intMapping = Mapping.inferMapping(client, intSchema);
-    assertNotNull(intMapping.get("properties").get("foo").get("null_value"));
-    assertEquals(0, intMapping.get("properties").get("foo").get("null_value").asInt());
-  }
-
-
-  protected Schema createSchema() {
-    Schema structSchema = createInnerSchema();
-    return SchemaBuilder.struct().name("record")
-        .field("boolean", Schema.BOOLEAN_SCHEMA)
-        .field("bytes", Schema.BYTES_SCHEMA)
-        .field("int8", Schema.INT8_SCHEMA)
-        .field("int16", Schema.INT16_SCHEMA)
-        .field("int32", Schema.INT32_SCHEMA)
-        .field("int64", Schema.INT64_SCHEMA)
-        .field("float32", Schema.FLOAT32_SCHEMA)
-        .field("float64", Schema.FLOAT64_SCHEMA)
-        .field("string", Schema.STRING_SCHEMA)
-        .field("array", SchemaBuilder.array(Schema.STRING_SCHEMA).build())
-        .field("map", SchemaBuilder.map(Schema.STRING_SCHEMA, Schema.STRING_SCHEMA).build())
-        .field("struct", structSchema)
-        .field("decimal", Decimal.schema(2))
-        .field("date", Date.SCHEMA)
-        .field("time", Time.SCHEMA)
-        .field("timestamp", Timestamp.SCHEMA)
-        .build();
-  }
-
-  private Schema createInnerSchema() {
-    return SchemaBuilder.struct().name("inner")
-        .field("boolean", Schema.BOOLEAN_SCHEMA)
-        .field("bytes", Schema.BYTES_SCHEMA)
-        .field("int8", Schema.INT8_SCHEMA)
-        .field("int16", Schema.INT16_SCHEMA)
-        .field("int32", Schema.INT32_SCHEMA)
-        .field("int64", Schema.INT64_SCHEMA)
-        .field("float32", Schema.FLOAT32_SCHEMA)
-        .field("float64", Schema.FLOAT64_SCHEMA)
-        .field("string", Schema.STRING_SCHEMA)
-        .field("array", SchemaBuilder.array(Schema.STRING_SCHEMA).build())
-        .field("map", SchemaBuilder.map(Schema.STRING_SCHEMA, Schema.STRING_SCHEMA).build())
-        .field("decimal", Decimal.schema(2))
-        .field("date", Date.SCHEMA)
-        .field("time", Time.SCHEMA)
-        .field("timestamp", Timestamp.SCHEMA)
-        .build();
-  }
-
-  @SuppressWarnings("unchecked")
-  private void verifyMapping(Schema schema, JsonObject mapping) throws Exception {
-    String schemaName = schema.name();
-    Object type = mapping.get("type");
-    if (schemaName != null) {
-      switch (schemaName) {
-        case Date.LOGICAL_NAME:
-        case Time.LOGICAL_NAME:
-        case Timestamp.LOGICAL_NAME:
-          assertEquals("\"" + ElasticsearchSinkConnectorConstants.DATE_TYPE + "\"", type.toString());
-          return;
-        case Decimal.LOGICAL_NAME:
-          assertEquals("\"" + ElasticsearchSinkConnectorConstants.DOUBLE_TYPE + "\"", type.toString());
-          return;
-      }
-    }
-
-    DataConverter converter = new DataConverter(true, BehaviorOnNullValues.IGNORE);
-    Schema.Type schemaType = schema.type();
-    switch (schemaType) {
-      case ARRAY:
-        verifyMapping(schema.valueSchema(), mapping);
-        break;
-      case MAP:
-        Schema newSchema = converter.preProcessSchema(schema);
-        JsonObject mapProperties = mapping.get("properties").getAsJsonObject();
-        verifyMapping(newSchema.keySchema(), mapProperties.get(ElasticsearchSinkConnectorConstants.MAP_KEY).getAsJsonObject());
-        verifyMapping(newSchema.valueSchema(), mapProperties.get(ElasticsearchSinkConnectorConstants.MAP_VALUE).getAsJsonObject());
-        break;
-      case STRUCT:
-        JsonObject properties = mapping.get("properties").getAsJsonObject();
-        for (Field field: schema.fields()) {
-          verifyMapping(field.schema(), properties.get(field.name()).getAsJsonObject());
-        }
-        break;
-      default:
-        assertEquals("\"" + Mapping.getElasticsearchType(client, schemaType) + "\"", type.toString());
-    }
-  }
-
-}
